[backend/frontend]  Fixed issue on number widget tooltip always being the same (#4071)
